# swagger.yaml
swagger: "2.0"
info:
  description: "Authors Haven - A social platform for the creative at heart."
  version: "1.0.0"
  title: "Authors Haven"
  license:
    name: "MIT"
    url: "https://www.apache.org/licenses/LICENSE-2.0.html"
basePath: "/api/v1"
...
schemes:
  - "https"
  - "http"
paths:
...
  /articles:
    post:
      tags:
      - "articles"
      summary: "Create article"
      description: "Creates a new article in the application"
      operationId: "createArticle"
      produces:
      - "application/json"
      - "application/xml"
      parameters:
      - in: "formData"
        name: "title"
        description: "The article title"
        required: true
        type: "string"
        format: "text"
        minLength: 4
        maxLength: 500
      - in: "formData"
        name: "body"
        description: "The article body"
        required: true
        type: "string"
        format: "text"
        minLength: 4
      - in: "formData"
        name: "image"
        description: "The image url"
        required: false
        type: "string"
        format: "text"
      responses:
        200:
          description: "Article created successfully"
          schema:
            $ref: "#/definitions/ApiResponse"
        400:
          description: "Title/ body is not allowed to be empty"
      security:
      - api_key: []
    get:
      tags:
      - "articles"
      summary: "User can get all articles"
      description: "Returns all available articles"
      operationId: "getArticles"
      produces:
      - "application/json"
      parameters: []
      responses:
        200:
          description: "success"
          schema:
            $ref: "#/definitions/ApiResponse"
        400:
          description: "fail"
  /articles/{slug}:
    get:
      tags:
      - "articles"
      summary: "User can get all articles"
      description: "Returns all available articles"
      operationId: "getArticle"
      produces:
      - "application/json"
      parameters:
      - name: "slug"
        in: "path"
        description: "Slug of the article to retrieve comments"
        required: true
        type: "string"
        format: "slug"
      responses:
        200:
          description: "success"
          schema:
            $ref: "#/definitions/ApiResponse"
        400:
          description: "fail"
    delete:
      tags:
      - "articles"
      summary: "User can delete an articles"
      description: "Deletes an article in application"
      operationId: "deleteArticles"
      produces:
      - "application/json"
      parameters: []
      responses:
        200:
          description: "success"
          schema:
            $ref: "#/definitions/ApiResponse"
        400:
          description: "fail"
      security:
      - api_key: []
    put:
      tags:
      - "articles"
      summary: "Update article"
      description: "Updates an article in the application"
      operationId: "updateArticle"
      produces:
      - "application/json"
      - "application/xml"
      parameters:
      - in: "formData"
        name: "title"
        description: "The article title"
        required: true
        type: "string"
        format: "text"
        minLength: 4
        maxLength: 500
      - in: "formData"
        name: "body"
        description: "The article body"
        required: true
        type: "string"
        format: "text"
        minLength: 4
      - in: "formData"
        name: "image"
        description: "The image url"
        required: false
        type: "string"
        format: "text"
      responses:
        200:
          description: "Article updated successfully"
          schema:
            $ref: "#/definitions/ApiResponse"
        400:
          description: "Title/ body is not allowed to be empty"
  /articles/{slug}/rate:
    post:
      tags:
      - "articles"
      summary: "Rate Article"
      description: "Rates an article"
      operationId: "rateArticle"
      produces:
      - "application/json"
      parameters:
      - name: "slug"
        in: "path"
        description: "Slug of the article to be rated"
        required: true
        type: "string"
        format: "slug"
      - name: "rate"
        in: "formData"
        description: "Rating you are giving to the article"
        required: true
        type: "integer"
        format: "int64"
      responses:
        200:
          description: "Rating has been updated"
          schema:
            $ref: "#/definitions/ApiResponse"
        201:
          description: "New rating has been made"
          schema:
            $ref: "#/definitions/ApiResponse"
        400:
          description: "Invalid rating supplied"
        404:
          description: "Article does not exist"
      security:
      - api_key: []
    get:
      tags:
      - "articles"
      summary: "Get ratings"
      description: "Get all article ratings"
      operationId: "getArticleRatings"
      produces:
      - "application/json"
      parameters:
      - name: "slug"
        in: "path"
        description: "Slug of the article to be get ratings"
        required: true
        type: "string"
        format: "slug"
      responses:
        200:
          description: "success"
          schema:
            $ref: "#/definitions/ApiResponse"
        400:
          description: "fail"
  /articles/{slug}/comments:
    post:
      tags:
      - "articles"
      summary: "Comment On Article"
      description: "Leaves a comment on an article"
      operationId: "create"
      produces:
      - "application/json"
      parameters:
      - name: "slug"
        in: "path"
        description: "Slug of the article to comment on"
        required: true
        type: "string"
        format: "slug"
      - name: "comment"
        in: "formData"
        description: "Comment you are giving the article"
        required: true
        type: "string"
        format: "text"
      responses:
        201:
          description: "Comment added successfully."
          schema:
            $ref: "#/definitions/ApiResponse"
        400:
          description: "Bad Request"
        404:
          description: "Article does not exist"
      security:
      - api_key: []
    get:
      tags:
      - "articles"
      summary: "Get comments"
      description: "Gets comments left on an article"
      operationId: "getComments"
      produces:
      - "application/json"
      parameters:
      - name: "slug"
        in: "path"
        description: "Slug of the article to retrieve comments"
        required: true
        type: "string"
        format: "slug"
      responses:
        200:
          description: "Comments retrieved successfully."
          schema:
            $ref: "#/definitions/ApiResponse"
        400:
          description: "Bad Request"
        404:
          description: "Article does not exist"
      security:
      - api_key: []
  /articles/{slug}/comments/{id}:
    put:
      tags:
      - "articles"
      summary: "Update Comment"
      description: "Updates a comment on an article"
      operationId: "updateComment"
      produces:
      - "application/json"
      parameters:
      - name: "slug"
        in: "path"
        description: "Slug of the article to update comment"
        required: true
        type: "string"
        format: "text"
      - name: "id"
        in: "path"
        description: "Id of the comment to be updated"
        required: true
        type: "integer"
        format: "int64"
      - name: "comment"
        in: "formData"
        description: "Comment you are updating"
        required: true
        type: "string"
        format: "text"
      responses:
        200:
          description: "Comment updated successfully."
          schema:
            $ref: "#/definitions/ApiResponse"
        400:
          description: "Bad Request"
        404:
          description: "Article does not exist"
      security:
      - api_key: []
    delete:
      tags:
      - "articles"
      summary: "Delete Comment"
      description: "Deletes a comment from an article"
      operationId: "deleteComment"
      produces:
      - "application/json"
      parameters:
      - name: "slug"
        in: "path"
        description: "Id of the article to delete comment"
        required: true
        type: "string"
        format: "text"
      - name: "id"
        in: "path"
        description: "Id of the comment to be deleted"
        required: true
        type: "integer"
        format: "int64"
      responses:
        200:
          description: "Comment deleted successfully."
          schema:
            $ref: "#/definitions/ApiResponse"
        400:
          description: "Bad Request"
        404:
          description: "Article does not exist"
      security:
      - api_key: []
  /articles/{slug}/comments/{id}/like:
    post:
      tags:
      - "articles"
      summary: "Like Comment"
      description: "Likes a comment"
      operationId: "likeComment"
      produces:
      - "application/json"
      parameters:
      - name: "slug"
        in: "path"
        description: "Slug of the article with the comment"
        required: true
        type: "string"
        format: "text"
      - name: "id"
        in: "path"
        description: "Id of the comment to be liked"
        required: true
        type: "integer"
        format: "int64"
      responses:
        200:
          description: "Comment liked successfully."
          schema:
            $ref: "#/definitions/ApiResponse"
        400:
          description: "Bad Request"
        404:
          description: "Comment does not exist"
      security:
      - api_key: []
    delete:
      tags:
      - "articles"
      summary: "Unlike Comment"
      description: "Unlikes a comment"
      operationId: "unlikeComment"
      produces:
      - "application/json"
      parameters:
      - name: "slug"
        in: "path"
        description: "Slug of the article to unlike comment"
        required: true
        type: "string"
        format: "text"
      - name: "id"
        in: "path"
        description: "Id of the comment to be unliked"
        required: true
        type: "integer"
        format: "int64"
      responses:
        200:
          description: "Comment unliked successfully."
          schema:
            $ref: "#/definitions/ApiResponse"
        400:
          description: "Bad Request"
        404:
          description: "Comment does not exist"
      security:
      - api_key: []
securityDefinitions:
  api_key:
    type: "apiKey"
    name: "Authorization"
    in: "header"
definitions:
...
  ApiResponse:
    type: "object"
    properties:
      status:
        type: "integer"
        format: "int32"
      data:
        type: "object"
        items:
          type: "object"
...